Split Kira wrapper into kira integration crate
