+++ crates/phonon-kira/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "phonon-kira"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+phonon = { path = "../phonon" }
+kira = "0.8"
