+++ crates/phonon/Cargo.toml
@@ -11,7 +11,6 @@ rand = "0.8"
 ndarray = "0.15"
 
 [dev-dependencies]
-kira = "0.8"
 criterion = { version = "0.5", features = ["html_reports"] }
 plotters = "0.3"
